[Upd] moved settings in src
