+++ manage.py
@@ -3,6 +3,6 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "src.settings")
     from django.core.management import execute_from_command_line
     execute_from_command_line(sys.argv)
